requirements.txt
# Core deep learning
torch>=2.0.0
torchvision>=0.15.0

# Image processing
Pillow>=9.0.0
opencv-python-headless>=4.5.5.64
scikit-image>=0.19.0
numpy>=1.23.0

# ML metrics
scikit-learn>=1.3.0

# FastAPI stack
fastapi>=0.103.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
httpx>=0.25.0

# MLflow
mlflow>=2.7.0

# Frontend
streamlit>=1.28.0
requests>=2.31.0

# Testing
pytest>=7.4.0

# Linting (dev)
flake8>=6.0.0
black>=23.0.0
isort>=5.12.0
